refactor : Concert 패키지 구조 리팩토링 및 안쓰는 함수 제거
