+++ src/main/java/com/example/book_your_seat/concert/controller/dto/ResultRedisConcert.java
@@ -1,21 +1,18 @@
 package com.example.book_your_seat.concert.controller.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.util.ArrayList;
 import java.util.List;
 
 
 @Getter
+@NoArgsConstructor
+@AllArgsConstructor
 public class ResultRedisConcert {
 
     private List<ConcertListResponse> concertList = new ArrayList<>();
-
-    public ResultRedisConcert(List<ConcertListResponse> concertList) {
-        this.concertList = concertList;
-    }
-
-    public ResultRedisConcert() {
-    }
 }
