--- src/main/java/com/example/book_your_seat/concert/service/ConcertCommandService.java
@@ -1,10 +1,0 @@
-package com.example.book_your_seat.concert.service;
-
-import com.example.book_your_seat.concert.controller.dto.AddConcertRequest;
-import com.example.book_your_seat.concert.controller.dto.ConcertResponse;
-
-public interface ConcertCommandService {
-
-    Long add(AddConcertRequest request);
-    void delete(Long id);
-}
