+++ src/main/java/com/example/book_your_seat/concert/service/query/ConcertQueryService.java
@@ -1,4 +1,4 @@
-package com.example.book_your_seat.concert.service;
+package com.example.book_your_seat.concert.service.query;
 
 import com.example.book_your_seat.concert.controller.dto.ConcertListResponse;
 import com.example.book_your_seat.concert.controller.dto.ConcertResponse;
@@ -13,37 +13,21 @@ import java.util.List;
 
 import static com.example.book_your_seat.concert.ConcertConst.INVALID_CONCERT_ID;
 
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
 @Service
-public class ConcertQueryServiceImpl implements ConcertQueryService {
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class ConcertQueryService {
 
     private final ConcertRepository concertRepository;
 
-    @Override
-    public List<ConcertResponse> findAll() {
-        return concertRepository.findAll().stream()
-                .map(ConcertResponse::from)
-                .toList();
-    }
-
-    @Override
     public ConcertResponse findById(final Long id) {
         return concertRepository.findById(id)
                 .map(ConcertResponse::from)
                 .orElseThrow(() -> new IllegalArgumentException(INVALID_CONCERT_ID + id));
     }
 
-    @Override
-    public List<ConcertListResponse> findAllConcertList() {
-        return concertRepository.findAll()
-                .stream().map(ConcertListResponse::new)
-                .toList();
-    }
-
-    @Override
     @Cacheable(cacheNames = "concerts", key = "'allConcerts'")
-    public ResultRedisConcert findUsedRedisAllConcertList() {
+    public ResultRedisConcert findAllConcertList() {
         List<ConcertListResponse> list = concertRepository.findAll()
                 .stream().map(ConcertListResponse::new)
                 .toList();
